[kernel][pmm] remove users of pmm_alloc_kpage
This is to help with an upcoming restructure of the way pages are
allocated from the pmm.
There wasn't much gain from using this convenience routine as it
currently stands.
Also remove the logic on the arm mmu code to allocate non page sized
page tables, which we currently dont support.

Change-Id: I7c834a7e3f4e8d2a3e73be8b092c9ccb9831b675